Started with a little refactoring as warmup for todays stage on the pattern matching tour
