+++ javaslang-pattern-matching-generator/src/main/java/javaslang/BaseChecker.java
@@ -1,0 +1,55 @@
+/*     / \____  _    _  ____   ______  / \ ____  __    _ _____
+ *    /  /    \/ \  / \/    \ /  /\__\/  //    \/  \  / /  _  \   Javaslang
+ *  _/  /  /\  \  \/  /  /\  \\__\\  \  //  /\  \ /\\/  \__/  /   Copyright 2014-now Daniel Dietrich
+ * /___/\_/  \_/\____/\_/  \_/\__\/__/___\_/  \_//  \__/_____/    Licensed under the Apache License, Version 2.0
+ */
+package javaslang;
+
+import javax.annotation.processing.Messager;
+import javax.lang.model.element.Element;
+import javax.lang.model.element.Modifier;
+import javax.tools.Diagnostic;
+import java.util.Arrays;
+import java.util.Set;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
+
+/**
+ * A base class for checking annotated {@link javax.lang.model.element.Element}s during annotation processing.
+ *
+ * @param <SELF> The class of this {@code BaseChecker} implementation.
+ * @param <E> The element type to be checked.
+ */
+abstract class BaseChecker<SELF extends BaseChecker<SELF, E>, E extends Element> {
+
+    final E elem;
+    final Messager messager;
+
+    BaseChecker(E elem, Messager messager) {
+        this.elem = elem;
+        this.messager = messager;
+    }
+
+    boolean ensure(Predicate<? super SELF> condition, Supplier<String> msg) {
+        @SuppressWarnings("unchecked")
+        final boolean result = condition.test((SELF) this);
+        if (!result) {
+            messager.printMessage(Diagnostic.Kind.ERROR, msg.get(), elem);
+        }
+        return result;
+    }
+
+    boolean hasAll(Modifier... modifiers) {
+        return elem.getModifiers().containsAll(Arrays.asList(modifiers));
+    }
+
+    boolean hasNone(Modifier... modifiers) {
+        final Set<Modifier> set = elem.getModifiers();
+        for (Modifier modifier : modifiers) {
+            if (set.contains(modifier)) {
+                return false;
+            }
+        }
+        return true;
+    }
+}
